+++ frontend/public/components/Dashboard.vue
@@ -8,36 +8,7 @@
         </div>
         <div v-if="currTab == 'myForms' && !formPreview" id="form-creation">
             <FieldAdd></FieldAdd>
-            <div id="form-container">
-                <div id="form-content" v-if="!formSaved">
-                    <input placeholder="Form Title" type="text" id="form-title" autocomplete="off" v-model="formTitle">
-                    <i class="fa fa-eye eye-icon" style="font-size:2rem; color:#2C666E;" @click="openFormPreview"></i>
-                    <i class="fa fa-save save-icon" style="font-size:2rem; color:#2C666E;" @click="saveForm"></i>
-                    <p v-if="fields.length == 0">Add some content to your form!</p>
-                    <div v-for="field in fields" class="field-card" v-bind:key="field.label"
-                    draggable 
-                    @dragstart='onDrag(field)' 
-                    @dragenter='onDragEnter($event)' 
-                    @dragover.prevent 
-                    @dragenter.prevent>
-                        <div class="field-card-inner">
-                            <input class="field-label" v-model="field.label" @input="resizeInput(field)">
-                            <span class="field-type">{{"(" + caps(field.type) + ")"}}</span>
-                            <i class="fa fa-bars clickable"></i>
-                            <i class="fa fa-times clickable" @click="deleteField(field)"></i>
-                            <ul v-for="option in field.options" v-bind:key="option">
-                                <li>{{option}}</li>
-                            </ul>
-                            <ul v-for="property in (field.slider ? Object.keys(field.slider) : [])" v-bind:key="field.label + property">
-                                <li>{{caps(property) + ": " + field.slider[property]}}</li>
-                            </ul>
-                        </div>
-                    </div>
-                </div>
-                <div v-if="formSaved">
-                    <p>Your form has been saved to the database.</p>
-                </div>
-            </div>
+            <FormCreation></FormCreation>
         </div>
         <FormPreview></FormPreview>
     </div>
@@ -51,6 +22,7 @@
     import Inbox from '../components/Inbox.vue';
     import TaskList from '../components/TaskList.vue';
     import FieldAdd from '../components/FieldAdd.vue';
+    import FormCreation from '../components/FormCreation.vue';
     import FormPreview from '../components/FormPreview.vue';
 
     export default {
@@ -60,6 +32,7 @@
             Inbox,
             TaskList,
             FieldAdd,
+            FormCreation,
             FormPreview
         },
         data () {
@@ -141,60 +114,6 @@
                 }
             }
             this.workload = workload;
-        },
-        saveForm() {
-            let form = {};
-            form.title = this.formTitle;
-            form.fields = this.fields.slice();
-            requests.postForm(form).then(
-                () => {
-                    this.formSaved = true;
-                }
-            );
-        },
-        openFormPreview() {
-            this.formPreview = true;
-        },
-        resizeInput(field) {
-            let fieldIndex = this.fields.indexOf(field);
-            let inputs = document.getElementsByClassName("field-label");
-            let currInput = inputs[fieldIndex];
-            currInput.style.width = (field.label.length * 0.95 + 2) + "ch";
-        },
-        deleteField(field) {
-            this.fields.splice(this.fields.indexOf(field),1);
-        },
-        onDrag(item) {
-            for (let i = 0; i < this.fields.length; i++)
-                if (this.fields[i] == item) {
-                    this.draggingPos = i;
-                    this.draggingField = this.fields[i];
-                    break;
-                }
-        },
-        onDragEnter(event) {
-            let targetInner, targetLabel, targetPos;
-            // Get the value of the input element we're dragging to
-            if (event.target.classList.contains("field-card"))
-                targetInner = event.target.getElementsByClassName("field-card-inner")[0];
-            else if (event.target.classList.contains("field-card-inner"))
-                targetInner = event.target;
-            else
-                targetInner = event.target.parentElement;
-            targetLabel = targetInner.getElementsByTagName("input")[0].value;
-
-            // Find it's position in fields
-            for (let i = 0; i < this.fields.length; i++)
-                if (this.fields[i].label == targetLabel) {
-                    targetPos = i;
-                    break;
-                }
-            
-            if (targetPos != this.draggingPos) {
-                this.fields.splice(this.draggingPos,1);
-                this.fields = this.fields.slice(0,targetPos).concat([this.draggingField]).concat(this.fields.slice(targetPos));
-                this.draggingPos = targetPos;
-            }
         },
         updateInbox() {
             requests.getUserForms("ann_perkins").then(
